Use UTF-8 for captured responses if charset is not set explicitly
In the newest versions of Spring Framework, after a change[1] to align
it with the RFC8259[2], the character encoding is no longer set to
UTF-8 for responses with `application/json` content-type.

This fix makes a `JsonResponseWrapper` use UTF-8 by default, instead of
getting charset from the wrapped `Response` instance which could have
been not set and so it returns default ISO-8895-1.

[1]: https://github.com/spring-projects/spring-framework/issues/22788
[2]: https://tools.ietf.org/html/rfc8259#section-11